requirements/local.txt
-r base.txt

Faker>=0.8
coverage>=4.5
django-nose>=1.4
django-debug-toolbar>=1.9
Whoosh>=2.7.4
prompt-toolkit>=1.0
ptpython>=0.41
